pyproject.toml
[project]
name = "f1-predictor"
version = "0.1.0"
description = "ML-powered F1 race outcome predictions using FastF1, XGBoost, and SHAP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<3.12"
authors = [
    { name = "Jowin Jestine", email = "jjestine@myolaris.com" },
]
dependencies = [
    "fastf1>=3.3",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "requests>=2.31",
    "numpy>=1.26",
    "google-cloud-storage>=2.16",
    "gcsfs>=2024.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pandas-stubs>=2.2",
    "types-requests>=2.31",
]
notebooks = [
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "scikit-learn>=1.4",
]
training = [
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "optuna>=3.6",
]
dl = [
    "rtdl-revisiting-models>=0.0.2",
]
serve = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic-settings>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/f1_predictor"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "TCH",
    "RUF",
]

[tool.ruff.lint.isort]
known-first-party = ["f1_predictor"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
packages = ["f1_predictor"]

[[tool.mypy.overrides]]
module = ["fastf1"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google.cloud.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["torch", "torch.*", "rtdl_revisiting_models"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sklearn.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["lightgbm", "xgboost"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pydantic_settings"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scipy", "scipy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "f1_predictor.models.dl_utils",
    "f1_predictor.models.architectures",
    "f1_predictor.models.sequence_architectures",
    "f1_predictor.models.quantile_architectures",
]
disallow_untyped_defs = false
warn_return_any = false
disable_error_code = ["misc", "no-any-return", "arg-type", "unused-ignore", "attr-defined", "type-arg", "assignment", "index"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=f1_predictor --cov-report=term-missing --cov-fail-under=30"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "nbconvert>=7.17.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]
